+++ Sources/KeyboardKit/App/KeyboardApp.swift
@@ -25,7 +25,7 @@ import SwiftUI
 ///             bundleId: "com.keyboardkit.demo",
 ///             appGroupId: "group.com.keyboardkit.demo",
 ///             locales: [.english, .swedish, .persian],
-///             dictationDeepLink: "keyboardkit://dictation"
+///             deepLinks: .init(app: "keyboardkit://")
 ///         )
 ///     }
 /// }
@@ -49,7 +49,34 @@ public struct KeyboardApp {
     ///   - keyboardExtensionBundleId: The app's keyboard bundle identifier, by default the bundle ID with a `.keyboard` suffix.
     ///   - appGroupId: The app's App Group identifier, if any.
     ///   - locales: The locales to use in the app, by default `.all`.
-    ///   - dictationDeepLink: The app's dictation deep link, if any.
+    ///   - deepLinks: App-specific deep links, if any.
+    public init(
+        name: String,
+        licenseKey: String = "",
+        bundleId: String,
+        keyboardExtensionBundleId: String? = nil,
+        appGroupId: String? = nil,
+        locales: [KeyboardLocale] = .all,
+        deepLinks: DeepLinks? = nil
+    ) {
+        self.name = name
+        self.bundleId = bundleId
+        self.keyboardExtensionBundleId = keyboardExtensionBundleId ?? "\(bundleId).keyboard"
+        self.appGroupId = appGroupId
+        self.locales = locales
+        self.licenseKey = licenseKey
+        self.deepLinks = deepLinks
+        if let appGroupId, let dictationLink = deepLinks?.dictation {
+            dictationConfiguration = .init(
+                appGroupId: appGroupId,
+                appDeepLink: dictationLink
+            )
+        } else {
+            dictationConfiguration = nil
+        }
+    }
+
+    @available(*, deprecated, message: "Use the deepLinks initializer instead.")
     public init(
         name: String,
         licenseKey: String = "",
@@ -59,20 +86,50 @@ public struct KeyboardApp {
         locales: [KeyboardLocale] = .all,
         dictationDeepLink: String? = ""
     ) {
-        self.name = name
-        self.bundleId = bundleId
-        self.keyboardExtensionBundleId = keyboardExtensionBundleId ?? "\(bundleId).keyboard"
-        self.appGroupId = appGroupId
-        self.locales = locales
-        self.licenseKey = licenseKey
-        if let appGroupId, let dictationDeepLink {
-            dictationConfiguration = .init(
-                appGroupId: appGroupId,
-                appDeepLink: dictationDeepLink
-            )
-        } else {
-            dictationConfiguration = nil
+        self.init(
+            name: name,
+            licenseKey: licenseKey,
+            bundleId: bundleId,
+            keyboardExtensionBundleId: keyboardExtensionBundleId,
+            appGroupId: appGroupId,
+            locales: locales,
+            deepLinks: .init(app: "", dictation: dictationDeepLink)
+        )
+    }
+
+    /// This type can define app-specific deep links.
+    public struct DeepLinks {
+
+        /// Create a custom keyboard deep links value.
+        ///
+        /// If you do not provide a value for a certain link,
+        /// a default link will be used to guide your design.
+        ///
+        /// - Parameters:
+        ///   - app: A deep link for opening the app, e.g. `x://`.
+        ///   - dictation: A deep link for opening the app and starting dictation, by default `x://dictation`.
+        ///   - keyboardSettings: A deep link for opening the app's keyboard settings screen, by default `x://keyboardSettings`.
+        ///   - languageSettings: A deep link for opening the app's language settings screen, by default `x://languageSettings`.
+        ///   - themeSettings: A deep link for opening the app's theme settings screen, by default `x://themeSettings`.
+        public init(
+            app: String,
+            dictation: String? = nil,
+            keyboardSettings: String? = nil,
+            languageSettings: String? = nil,
+            themeSettings: String? = nil
+        ) {
+            self.app = app
+            self.dictation = dictation ?? "\(app)/dictation"
+            self.keyboardSettings = keyboardSettings ?? "\(app)/keyboardSettings"
+            self.languageSettings = languageSettings ?? "\(app)/languageSettings"
+            self.themeSettings = themeSettings ?? "\(app)/themeSettings"
         }
+
+        public let app: String
+        public let dictation: String
+        public let keyboardSettings: String
+        public let languageSettings: String
+        public let themeSettings: String
     }
 
     /// The name of the app.
@@ -93,7 +150,10 @@ public struct KeyboardApp {
     /// The locales to use in the app.
     public let locales: [KeyboardLocale]
 
-    /// The app's dictation deep link, if any.
+    /// App-specific deep links, if any.
+    public let deepLinks: DeepLinks?
+
+    /// The app's dictation configuration, if any.
     public let dictationConfiguration: Dictation.KeyboardConfiguration?
 }
 
@@ -116,7 +176,7 @@ private extension KeyboardApp {
             keyboardExtensionBundleId: "com.keyboardkit.demo.keyboard",
             appGroupId: "group.com.keyboardkit.demo",
             locales: [.english, .swedish, .persian],
-            dictationDeepLink: "keyboardkit://dictation"
+            deepLinks: .init(app: "keyboardkit://")
         )
     }
 }